add reference StoreUnit to Drug

+++ db/migrate/20160621161207_create_drugs.rb
@@ -4,7 +4,7 @@ class CreateDrugs < ActiveRecord::Migration
       t.text :name
       t.text :trade_name
       t.text :effect
-      t.decimal :balance
+      t.integer :balance
       t.references :drug_usage, index: true, foreign_key: true
 
       t.timestamps null: false
